backend/requirements.txt
fastapi
uvicorn
python-socketio
pymodbus
python-dotenv